requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.3.1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
daphne==3.0.2
dj-database-url==2.1.0
Django==4.2.11
django-cors-headers==4.3.1
django-rest-knox==4.2.0
djangorestframework==3.15.1
google-api-core==2.19.1
google-api-python-client==2.88.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.63.2
gunicorn==22.0.0
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
iniconfig==2.0.0
lxml==5.3.0
oauthlib==3.2.2
packaging==24.0
-e git+https://github.com/rigomoreirar/pass_gen.git@aef7dba85c9251294da24221fc568225084b7e4a#egg=password_manager
pillow==10.4.0
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.1
pycparser==2.22
PyMuPDF==1.24.12
pyOpenSSL==24.2.1
pyparsing==3.1.2
pytest==8.3.3
python-dotenv==1.0.1
python-http-client==3.3.7
python-pptx==1.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sendgrid==3.0.0
service-identity==24.1.0
setuptools==71.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.19
uvicorn==0.30.5
XlsxWriter==3.2.0
zope.interface==6.4.post2
